// UserRegistrationApp/src/app/models/user.model.ts
export interface User {
    id?: number;
    fullName: string;
    profileImage: string;
    gender: string;
    address: string;
    city: string;
    pin: number;
    state: string;
    country: string;
    email: string;
    contact: string;
    educationQualification: string;
    designation: string;
  }